src/spark-project/spark-common/src/main/java/io/kyligence/kap/spark/common/logging/PrintStreamManager.java
/*
 * Copyright (C) 2016 Kyligence Inc. All rights reserved.
 *
 * http://kyligence.io
 *
 * This software is the confidential and proprietary information of
 * Kyligence Inc. ("Confidential Information"). You shall not disclose
 * such Confidential Information and shall use it only in accordance
 * with the terms of the license agreement you entered into with
 * Kyligence Inc.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
 * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
 * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
 * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
 * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
 * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
 * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
 * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
package io.kyligence.kap.spark.common.logging;


import org.slf4j.Logger;

import java.io.PrintStream;

public class PrintStreamManager {

    private PrintStreamManager() {
    }

    private static final PrintStream outPrintStream = System.out;
    private static final PrintStream errPrintStream = System.err;

    public static PrintStream createLoggingProxy(final Logger logger, final PrintStream printStream, boolean setError) {
        return new PrintStream(printStream) {
            @Override
            public void print(final String message) {
                if (setError) {
                    logger.error(message);
                } else {
                    logger.info(message);
                }
            }

            @Override
            public void println(final String message) {
                print(message);
            }
        };
    }

    public static void resetSystemPrintStream() {
        System.setOut(outPrintStream);
        System.setErr(errPrintStream);
    }

}
